requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
sqlmodel==0.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.24.1
psycopg2==2.9.9
python-multipart==0.0.6
alembic==1.11.1
